Update wrapper to support the status command. Remove the restart command.

+++ ses-server-wrapper.py
@@ -17,19 +17,16 @@ def run(command):
     python_file = 'ses-server.py'
     pid_file = os.path.join(CACHE_FOLDER, "pid-" + str(port))
 
-    if command == "restart":
-        kill_process(pid_file)
-        runPythonScript(pid_file, port, python_file)
-    elif command == "start":
-        runPythonScript(pid_file, port, python_file)
-    elif command == "stop":
-        kill_process(pid_file)
+    if command == "start":
+        run_python(python_file, port, pid_file)
+        wait_until_port_is_open(port, 5, 5)
 
+    if command == "status":
+        wait_until_port_is_open(port, 1, 0)
 
-def runPythonScript(pid_file, port, python_file):
-    pid = run_python(python_file, port, pid_file)
-    wait_until_port_is_open(port, FIVE_SECONDS_DELAY)
-    print "Process running as pid: " + str(pid)
+    if command == "stop":
+        kill_process(pid_file)
+        wait_until_port_is_closed(port, 5, 5)
 
 
 def run_python(python_path, port, pid_file):
@@ -37,36 +34,61 @@ def run_python(python_path, port, pid_file):
     f = open(pid_file, "w")
     f.write(str(proc.pid))
     f.close()
+    print "Process running as pid: " + str(proc.pid)
     return proc.pid
 
 
-def wait_until_port_is_open(port, delay):
+def wait_until_port_is_open(port, count, delay):
     n = 0
-    while n < delay:
+    while True:
         print "Is application listening on port " + str(port) + "? "
         sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
         result = sock.connect_ex(('127.0.0.1', port))
         if result == 0:
             print "Yes"
             return
-        print "No. Retrying in " + str(delay) + " seconds"
+
         n = n + 1
-        time.sleep(delay)
+        if n < count:
+            print "No. Retrying in " + str(delay) + " seconds"
+            time.sleep(delay)
+        else:
+            print "No."
+            return
+
+
+def wait_until_port_is_closed(port, count, delay):
+    n = 0
+    while True:
+        print "Is application listening on port " + str(port) + "? "
+        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+        result = sock.connect_ex(('127.0.0.1', int(port)))
+        if result != 0:
+            print "No"
+            return
+
+        n = n + 1
+        if n < count:
+            print "Yes. Retrying in " + str(delay) + " seconds"
+            time.sleep(delay)
+        else:
+            print "Yes."
+            return
 
 
 def kill_process(pid_file):
+    if not os.path.exists(pid_file):
+        print "Already stopped."
+        return
+
+    f = open(pid_file, "r")
     try:
-        f = open(pid_file, "r")
-        try:
-            pid_str = f.read()
-            print "Kill process with pid: " + pid_str
-            os.kill(int(pid_str), signal.SIGTERM)
-        except Exception:
-            f.close()
-            os.remove(pid_file)
-    except IOError as error:
-        print "Error: {0}, when trying to open {1}...".format(error, pid_file)
-        print "The process never existed or may not exist anymore, let's continue..."
+        pid_str = f.read()
+        print "Kill process with pid: " + pid_str
+        os.kill(int(pid_str), signal.SIGTERM)
+    except Exception:
+        f.close()
+        os.remove(pid_file)
 
 
 if __name__ == "__main__":
